Moving logic to State
A state is the vital component now.
Every frame of a game is a state and an action (either a move or a drop)
creates a new frame. A state contains all information of the Hive, the bee move
and the turn number. The logic of the insects was moved to either the hive
or the state itself.